+++ python/taco/__init__.py
@@ -4,7 +4,7 @@ from importlib.metadata import PackageNotFoundError, version
 
 from . import metadata, reader
 from .contract import Asset, Collection, Contract, Folder, Sample
-from .dataset import Dataset, open
+from .dataset import Dataset, open_dataset, read
 from .errors import TacoError
 from .schema import CollectionMetadata, DerivedMetadata, Level, Metadata, MetadataSchema
 from .tacocat import consolidate
@@ -32,8 +32,9 @@ __all__ = [
     "__version__",
     "consolidate",
     "metadata",
-    "open",
+    "open_dataset",
     "open_writer",
+    "read",
     "reader",
     "validate",
 ]